refactor: view 계층 분리
